Move FeatureTest methods to trait
